google_secops_threat_hunting/requirements.txt
google-cloud-securitycenter>=1.10.0
google-cloud-logging>=3.0.0
pandas>=1.3.0
jupyter>=1.0.0
notebook>=6.4.0
python-dotenv>=0.19.0
requests>=2.26.0
pyyaml>=5.4.1 